requirements.txt
# Dependencias principales
PyYAML>=6.0

# Dependencias de desarrollo y testing
pytest>=8.0

pytest-cov>=4.1.0